requirements.txt
Django==5.0.1
django-crispy-forms==2.1
django-bootstrap4==25.0
Pillow==10.1.0
python-dotenv==1.0.0
